Move app store & reducer in own directory
