+++ modules/hitl2/src/views/full/Components/Sidebar.tsx
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 import React, { FC, useContext, useState } from 'react'
 
-import { Context } from '../Store'
+import { Context } from '../app/Store'
 
 import { ApiType } from './../Api'
 import { EscalationType } from '../../../types'
